+++ server/.gitignore
@@ -108,10 +108,12 @@ dist
 
 #Upload of files
 tmp/*
+images/*
 !tmp/.gitkeep
 
 #Configs
 config/
+.vscode
 
 #Database
 *.sqlite
