+++ server/src/shared/errors/AppError.ts
