+++ server/src/shared/middlewares/ensureAuthenticate.ts
@@ -2,7 +2,7 @@ import { Request, Response, NextFunction } from 'express';
 import { verify } from 'jsonwebtoken';
 
 import { authConfig } from '@config/auth';
-import { AppError } from 'errors/AppError';
+import { AppError } from '@shared/errors/AppError';
 
 interface TokenPayload {
   iat: number;
