+++ server/tsconfig.json
@@ -33,6 +33,7 @@
       "@controllers/*": ["controllers/*"],
       "@config/*": ["config/*"],
       "@models/*": ["models/*"],
+      "@shared/*": ["shared/*"],
     },                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
